добавлены разрешения для пользователей

+++ app_item/app/endpoints/item_router.py
@@ -63,6 +63,16 @@ item_router = APIRouter(prefix='/item', tags=['Item'])
 target_service_url = "http://192.168.1.92:86"
 
 
+def user_staff_admin(role):
+    if role == "client" or role == "staff" or role == "admin":
+        return True
+    return False
+
+def staff_admin(role):
+    if role == "staff" or role == "admin":
+        return True
+    return False
+
 def make_request_to_target_service(item_id, size, count, price, name, user):
     url = f"{target_service_url}/api/cart/"
     headers = {
@@ -76,13 +86,6 @@ def make_request_to_target_service(item_id, size, count, price, name, user):
         return response.status_code
     else:
         raise Exception(f"Error making request:{response.status_code}, {response.text}")
-
-
-# class dropdownChoices(str, Enum):
-#     xs = "xs"
-#     s = "s"
-#     m = "m"
-#     l = "l"
 
 
 @item_router.get('/')
@@ -112,13 +115,18 @@ def create_item(
         name: str,
         price: float,
         design: str,
-        item_service: ItemService = Depends(ItemService)) -> Item:
+        item_service: ItemService = Depends(ItemService),
+        user: str = Header(...)) -> Item:
+    user = eval(user)
     with tracer.start_as_current_span("Create item") as span:
         add_endpoint_info(span, "/ [POST]")
         try:
-            item = item_service.create_item(name, price, design)
-            add_operation_result(span, "success")
-            return item.dict()
+            if user['id'] is not None:
+                if staff_admin(user['role']):
+                    item = item_service.create_item(name, price, design)
+                    add_operation_result(span, "success")
+                    return item.dict()
+                raise HTTPException(403)
         except KeyError:
             add_operation_result(span, "failure: Cant create item")
             raise HTTPException(404, f'Cant create item')
@@ -135,10 +143,13 @@ def add_to_cart(
     with tracer.start_as_current_span("Add to cart") as span:
         add_endpoint_info(span, "/add_to_cart")
         try:
-            item = item_service.get_items_by_id(item_id)
-            make_request_to_target_service(item_id, size, count, item.price, item.name, user)
-            add_operation_result(span, "success")
-            return item
+            if user['id'] is not None:
+                if user_staff_admin(user['role']):
+                    item = item_service.get_items_by_id(item_id)
+                    make_request_to_target_service(item_id, size, count, item.price, item.name, user)
+                    add_operation_result(span, "success")
+                    return item
+                raise HTTPException(403)
         except KeyError:
             add_operation_result(span, "failure: Cant add to cart item")
             raise HTTPException(404, f'Cant add to cart item')
@@ -179,17 +190,22 @@ def delete_design(
         except KeyError:
             add_operation_result(span, "failure")
             raise HTTPException(404, f'Design with id={id} not found')
-        
+
+
 @item_router.delete('/delete_item/{id}')
 def delete_item(id: UUID, 
                 item_service: ItemService = Depends(ItemService),
                 user: str = Header(...),) -> Item:
+    user = eval(user)
     with tracer.start_as_current_span("Delete design") as span:
         add_endpoint_info(span, "/delete-design/{id}")
         try:
-            item = item_service.delete_design(id)
-            add_operation_result(span, "success")
-            return item.dict()
+            if user['id'] is not None:
+                if staff_admin(user['role']):
+                    item = item_service.delete_design(id)
+                    add_operation_result(span, "success")
+                    return item.dict()
+                raise HTTPException(403)
         except KeyError:
             add_operation_result(span, "failure")
             raise HTTPException(404, f'Design with id={id} not found')
